Add files via upload
Project result:
Summary
So this is how we can analyze the Uber trips for New York City. Some of the conclusions that I got from this analysis are:
Monday is the most profitable day for Uber On Saturdays less number of people use Uber 6 pm is the busiest day for Uber On average a rise in Uber trips starts around 5 am. Most of the Uber trips originate near the Manhattan region in New York.